norm ing...

+++ libft/ft_strlcpy.c
@@ -6,15 +6,15 @@
 /*   By: gyeon <gyeon@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/10 17:18:21 by gyeon             #+#    #+#             */
-/*   Updated: 2021/05/10 17:18:24 by gyeon            ###   ########.fr       */
+/*   Updated: 2021/05/11 17:57:09 by gyeon            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t ft_strlcpy(char *restrict dst, const char *restrict src, size_t dstsize)
+size_t	ft_strlcpy(char *restrict dst, const char *restrict src, size_t dstsize)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (*(src + i) && (i + 1 < dstsize))
